requirements.txt
fastapi
uvicorn
python-multipart
tiktoken
PyPDF2
python-docx